+++ extensions/github/src/helpers/repository.ts
@@ -34,88 +34,14 @@ export const WEB_IDES = [
   },
 ];
 
-import { exec } from "node:child_process";
-import { existsSync, mkdirSync } from "node:fs";
-import path from "node:path";
-import { promisify } from "node:util";
-
-import { LocalStorage, Toast, getPreferenceValues, showToast } from "@raycast/api";
+import { LocalStorage } from "@raycast/api";
 import { useCachedState } from "@raycast/utils";
 import { useEffect } from "react";
 
 import { ExtendedRepositoryFieldsFragment } from "../generated/graphql";
 
-import { getErrorMessage } from "./errors";
-
 const VISITED_REPOSITORIES_KEY = "VISITED_REPOSITORIES";
 const VISITED_REPOSITORIES_LENGTH = 25;
-
-const execAsync = promisify(exec);
-
-export async function cloneAndOpen(repository: ExtendedRepositoryFieldsFragment, clonePath: string, branch?: string) {
-  const { application } = getPreferenceValues<Preferences.SearchRepositories>();
-  const applicationPath = application?.path.replaceAll(" ", "\\ ");
-
-  const repoName = repository.name;
-  let targetDir = path.join(clonePath, repoName);
-  let counter = 1;
-
-  while (existsSync(targetDir)) {
-    targetDir = path.join(clonePath, `${repoName}-${counter}`);
-    counter++;
-  }
-
-  try {
-    mkdirSync(targetDir, { recursive: true });
-  } catch (error) {
-    await showToast({
-      style: Toast.Style.Failure,
-      title: "Error creating directory",
-      message: getErrorMessage(error),
-    });
-    console.error(error);
-    return;
-  }
-
-  const openCommand = `open -a ${applicationPath} ${targetDir}`;
-
-  const toast = await showToast({
-    title: `Cloning ${repository.nameWithOwner}`,
-    message: `to ${targetDir}`,
-    style: Toast.Style.Animated,
-  });
-
-  const cloneUrl = `https://github.com/${repository.nameWithOwner}`;
-  const branchOption = branch ? `-b ${branch}` : "";
-  const cloneCommand = `git clone ${branchOption} ${cloneUrl} ${targetDir}`;
-
-  try {
-    await execAsync(cloneCommand);
-    toast.title = "Repository cloned successfully";
-    toast.style = Toast.Style.Success;
-  } catch (error) {
-    toast.style = Toast.Style.Failure;
-    toast.title = "Error while cloning the repository";
-    toast.message = getErrorMessage(error);
-    console.error(error);
-    return;
-  }
-
-  await new Promise((resolve) => setTimeout(resolve, 1000));
-
-  toast.title = "Opening repository...";
-
-  try {
-    await execAsync(openCommand);
-    toast.title = "Code editor launched!";
-    toast.style = Toast.Style.Success;
-  } catch (error) {
-    toast.style = Toast.Style.Failure;
-    toast.title = "Error while opening the repository";
-    toast.message = getErrorMessage(error);
-    console.error(error);
-  }
-}
 
 // History was stored in `LocalStorage` before, after migration it's stored in `Cache`
 async function loadVisitedRepositories() {